Add case search endpoint view integration tests
Also moves views.py content into the views/ package __init__.py to fix
import compatibility after the views module was converted to a package.
